refactor(meeting_facilitator/handoff): extract tests to siblings (#1266)
- handoff.rs: 984 -> 417 LOC
- new tests_handoff.rs (353 LOC) and tests_handoff_extra.rs (260 LOC)
- widen is_open_question and find_newest_handoff to pub(super)
