+++ src/meeting_facilitator/handoff.rs
@@ -85,7 +85,7 @@ fn is_rhetorical(note: &str) -> bool {
 const OPEN_QUESTION_PREFIXES: &[&str] = &["open:", "todo:", "question:", "tbd:", "unresolved:"];
 
 /// Returns `true` if `note` should be extracted as an open question.
-fn is_open_question(note: &str) -> bool {
+pub(super) fn is_open_question(note: &str) -> bool {
     let lower = note.trim().to_lowercase();
 
     // Explicit markers always count.
@@ -257,7 +257,7 @@ pub fn write_meeting_handoff(dir: &Path, handoff: &MeetingHandoff) -> SimardResu
 
 /// Find the newest handoff file in a directory (timestamped `handoff-*.json`
 /// or legacy `meeting_handoff.json`). Returns `None` if no file exists.
-fn find_newest_handoff(dir: &Path) -> Option<PathBuf> {
+pub(super) fn find_newest_handoff(dir: &Path) -> Option<PathBuf> {
     let mut candidates: Vec<PathBuf> = Vec::new();
 
     // Legacy fixed filename.
@@ -413,572 +413,4 @@ pub fn remove_session_wip(dir: &Path) -> SimardResult<()> {
         })?;
     }
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::meeting_facilitator::types::{ActionItem, MeetingDecision, MeetingSessionStatus};
-
-    /// Build a minimal session for testing.
-    fn make_session(
-        topic: &str,
-        notes: Vec<&str>,
-        decisions: Vec<MeetingDecision>,
-        action_items: Vec<ActionItem>,
-        participants: Vec<&str>,
-    ) -> MeetingSession {
-        MeetingSession {
-            topic: topic.to_string(),
-            decisions,
-            action_items,
-            notes: notes.into_iter().map(String::from).collect(),
-            status: MeetingSessionStatus::Closed,
-            started_at: chrono::Utc::now().to_rfc3339(),
-            participants: participants.into_iter().map(String::from).collect(),
-            explicit_questions: Vec::new(),
-            themes: Vec::new(),
-        }
-    }
-
-    fn sample_decision() -> MeetingDecision {
-        MeetingDecision {
-            description: "Ship phase 8".to_string(),
-            rationale: "Unblocks goal curation".to_string(),
-            participants: vec!["alice".to_string()],
-        }
-    }
-
-    fn sample_action() -> ActionItem {
-        ActionItem {
-            description: "Write tests".to_string(),
-            owner: "bob".to_string(),
-            priority: 1,
-            due_description: Some("end of sprint".to_string()),
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // is_open_question / is_rhetorical unit tests
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn genuine_question_is_extracted() {
-        assert!(is_open_question("What is the timeline for phase 9?"));
-        assert!(is_open_question(
-            "How should we handle backward compatibility?"
-        ));
-    }
-
-    #[test]
-    fn rhetorical_short_question_is_filtered() {
-        assert!(!is_open_question("Why not?"));
-        assert!(!is_open_question("Right?"));
-        assert!(!is_open_question("Isn't it?"));
-        assert!(!is_open_question("So what?"));
-    }
-
-    #[test]
-    fn rhetorical_trailing_pattern_is_filtered() {
-        assert!(!is_open_question(
-            "We should deploy on Monday, don't you think?"
-        ));
-        assert!(!is_open_question("The fix looks good, right?"));
-    }
-
-    #[test]
-    fn explicit_markers_without_question_mark() {
-        assert!(is_open_question("OPEN: decide on migration strategy"));
-        assert!(is_open_question("todo: finalize API contract"));
-        assert!(is_open_question("Question: ownership of the rollback plan"));
-        assert!(is_open_question("TBD: release date"));
-        assert!(is_open_question(
-            "UNRESOLVED: cross-team dependency on auth service"
-        ));
-    }
-
-    #[test]
-    fn plain_note_without_marker_or_question_mark_is_ignored() {
-        assert!(!is_open_question("We agreed to use Postgres"));
-        assert!(!is_open_question("Deployment target is Friday"));
-    }
-
-    // -----------------------------------------------------------------------
-    // MeetingHandoff::from_session
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn from_session_populates_all_fields() {
-        let session = make_session(
-            "Sprint planning",
-            vec!["note 1", "What is the timeline for phase 9?"],
-            vec![sample_decision()],
-            vec![sample_action()],
-            vec!["alice", "bob"],
-        );
-
-        let handoff = MeetingHandoff::from_session(&session);
-
-        assert_eq!(handoff.topic, "Sprint planning");
-        assert!(!handoff.started_at.is_empty());
-        assert!(!handoff.closed_at.is_empty());
-        assert_eq!(handoff.decisions.len(), 1);
-        assert_eq!(handoff.action_items.len(), 1);
-        assert_eq!(
-            handoff.open_questions,
-            vec![OpenQuestion {
-                text: "What is the timeline for phase 9?".to_string(),
-                explicit: false,
-            }]
-        );
-        assert!(!handoff.processed);
-        assert!(handoff.duration_secs.is_some());
-        assert_eq!(handoff.transcript.len(), 2);
-        // alice (session) + bob (session) — alice is already in session.participants
-        // alice also appears in decision.participants but should not be duplicated.
-        assert!(handoff.participants.contains(&"alice".to_string()));
-        assert!(handoff.participants.contains(&"bob".to_string()));
-    }
-
-    #[test]
-    fn from_session_collects_unique_participants() {
-        let session = make_session(
-            "Dedup check",
-            vec![],
-            vec![MeetingDecision {
-                description: "d".to_string(),
-                rationale: "r".to_string(),
-                participants: vec!["alice".to_string(), "charlie".to_string()],
-            }],
-            vec![ActionItem {
-                description: "a".to_string(),
-                owner: "alice".to_string(),
-                priority: 1,
-                due_description: None,
-            }],
-            vec!["alice", "bob"],
-        );
-
-        let handoff = MeetingHandoff::from_session(&session);
-        // alice appears in session, decision, and action but should appear once.
-        assert_eq!(
-            handoff
-                .participants
-                .iter()
-                .filter(|p| *p == "alice")
-                .count(),
-            1
-        );
-        // charlie from the decision participant list should be added.
-        assert!(handoff.participants.contains(&"charlie".to_string()));
-    }
-
-    #[test]
-    fn from_session_empty_session() {
-        let session = make_session("Empty", vec![], vec![], vec![], vec![]);
-        let handoff = MeetingHandoff::from_session(&session);
-
-        assert_eq!(handoff.topic, "Empty");
-        assert!(handoff.decisions.is_empty());
-        assert!(handoff.action_items.is_empty());
-        assert!(handoff.open_questions.is_empty());
-        assert!(handoff.transcript.is_empty());
-        assert!(handoff.participants.is_empty());
-    }
-
-    #[test]
-    fn from_session_only_rhetorical_questions() {
-        let session = make_session(
-            "Rhetorical",
-            vec![
-                "Why not?",
-                "Right?",
-                "Looks good, don't you think?",
-                "Plain note without question",
-            ],
-            vec![],
-            vec![],
-            vec![],
-        );
-        let handoff = MeetingHandoff::from_session(&session);
-        assert!(
-            handoff.open_questions.is_empty(),
-            "Rhetorical questions should not appear in open_questions: {:?}",
-            handoff.open_questions
-        );
-    }
-
-    #[test]
-    fn from_session_explicit_markers_no_question_mark() {
-        let session = make_session(
-            "Markers",
-            vec![
-                "OPEN: decide on migration strategy",
-                "TODO: finalize API contract",
-                "Regular note",
-            ],
-            vec![],
-            vec![],
-            vec![],
-        );
-        let handoff = MeetingHandoff::from_session(&session);
-        assert_eq!(handoff.open_questions.len(), 2);
-        assert!(handoff.open_questions[0].text.starts_with("OPEN:"));
-        assert!(!handoff.open_questions[0].explicit);
-        assert!(handoff.open_questions[1].text.starts_with("TODO:"));
-        assert!(!handoff.open_questions[1].explicit);
-    }
-
-    #[test]
-    fn from_session_populates_themes_from_notes() {
-        let session = make_session(
-            "Theme test",
-            vec![
-                "We discussed testing strategies.",
-                "Testing coverage needs improvement.",
-                "More testing will help quality.",
-                "Deployment pipelines look good.",
-            ],
-            vec![],
-            vec![],
-            vec![],
-        );
-        let handoff = MeetingHandoff::from_session(&session);
-        assert!(
-            handoff.themes.contains(&"testing".to_string()),
-            "Expected 'testing' theme from recurring notes: {:?}",
-            handoff.themes
-        );
-    }
-
-    #[test]
-    fn from_session_empty_notes_no_themes() {
-        let session = make_session("No themes", vec![], vec![], vec![], vec![]);
-        let handoff = MeetingHandoff::from_session(&session);
-        assert!(handoff.themes.is_empty());
-    }
-
-    #[test]
-    fn from_session_explicit_themes_come_first() {
-        let mut session = make_session(
-            "Explicit theme test",
-            vec![
-                "We discussed testing strategies.",
-                "Testing coverage needs improvement.",
-                "More testing will help quality.",
-            ],
-            vec![],
-            vec![],
-            vec![],
-        );
-        session.themes = vec!["performance".to_string(), "reliability".to_string()];
-        let handoff = MeetingHandoff::from_session(&session);
-        // Explicit themes must appear before inferred ones
-        assert_eq!(handoff.themes[0], "performance");
-        assert_eq!(handoff.themes[1], "reliability");
-        // Inferred "testing" still present (not a duplicate)
-        assert!(
-            handoff.themes.contains(&"testing".to_string()),
-            "inferred theme should also appear: {:?}",
-            handoff.themes
-        );
-    }
-
-    #[test]
-    fn from_session_explicit_themes_deduplicated() {
-        let mut session = make_session("Dedup test", vec![], vec![], vec![], vec![]);
-        session.themes = vec!["Performance".to_string()];
-        // Inferred would also produce "performance" if it appeared in notes
-        // Just verify no duplicate casing issues in round-trip
-        let handoff = MeetingHandoff::from_session(&session);
-        let count = handoff
-            .themes
-            .iter()
-            .filter(|t| t.to_lowercase() == "performance")
-            .count();
-        assert_eq!(
-            count, 1,
-            "no duplicate performance theme: {:?}",
-            handoff.themes
-        );
-    }
-
-    // -----------------------------------------------------------------------
-    // write / load / mark_processed round-trip (filesystem)
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn write_and_load_round_trip() {
-        let dir = tempfile::tempdir().unwrap();
-        let session = make_session(
-            "Roundtrip",
-            vec!["What is the deadline for the release?"],
-            vec![sample_decision()],
-            vec![sample_action()],
-            vec!["alice"],
-        );
-        let handoff = MeetingHandoff::from_session(&session);
-        write_meeting_handoff(dir.path(), &handoff).unwrap();
-
-        let loaded = load_meeting_handoff(dir.path()).unwrap().unwrap();
-        assert_eq!(loaded.topic, "Roundtrip");
-        assert_eq!(loaded.decisions.len(), 1);
-        assert_eq!(loaded.action_items.len(), 1);
-        assert_eq!(loaded.open_questions.len(), 1);
-        assert!(!loaded.processed);
-    }
-
-    #[test]
-    fn load_from_empty_dir_returns_none() {
-        let dir = tempfile::tempdir().unwrap();
-        let result = load_meeting_handoff(dir.path()).unwrap();
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn mark_processed_persists() {
-        let dir = tempfile::tempdir().unwrap();
-        let session = make_session("Mark test", vec![], vec![], vec![], vec![]);
-        let handoff = MeetingHandoff::from_session(&session);
-        write_meeting_handoff(dir.path(), &handoff).unwrap();
-
-        mark_meeting_handoff_processed(dir.path()).unwrap();
-
-        let loaded = load_meeting_handoff(dir.path()).unwrap().unwrap();
-        assert!(loaded.processed);
-    }
-
-    #[test]
-    fn mark_processed_noop_on_empty_dir() {
-        let dir = tempfile::tempdir().unwrap();
-        // Should succeed without error even when no handoff exists.
-        mark_meeting_handoff_processed(dir.path()).unwrap();
-    }
-
-    #[test]
-    fn mark_processed_in_place_persists() {
-        let dir = tempfile::tempdir().unwrap();
-        let session = make_session("In-place", vec![], vec![], vec![], vec![]);
-        let mut handoff = MeetingHandoff::from_session(&session);
-        write_meeting_handoff(dir.path(), &handoff).unwrap();
-
-        mark_handoff_processed_in_place(dir.path(), &mut handoff).unwrap();
-        assert!(handoff.processed);
-
-        let loaded = load_meeting_handoff(dir.path()).unwrap().unwrap();
-        assert!(loaded.processed);
-    }
-
-    #[test]
-    fn serialization_round_trip_via_serde() {
-        let session = make_session(
-            "Serde test",
-            vec!["TBD: release date", "How do we handle rollback?"],
-            vec![sample_decision()],
-            vec![sample_action()],
-            vec!["alice"],
-        );
-        let handoff = MeetingHandoff::from_session(&session);
-
-        let json = serde_json::to_string_pretty(&handoff).unwrap();
-        let deser: MeetingHandoff = serde_json::from_str(&json).unwrap();
-        assert_eq!(handoff, deser);
-    }
-
-    #[test]
-    fn newest_handoff_wins_when_multiple_exist() {
-        let dir = tempfile::tempdir().unwrap();
-        // Write two handoffs with different topics — the second has a later
-        // timestamp so it should be the one returned by load.
-        let s1 = make_session("First", vec![], vec![], vec![], vec![]);
-        let h1 = MeetingHandoff::from_session(&s1);
-        write_meeting_handoff(dir.path(), &h1).unwrap();
-
-        // Small sleep so the timestamps differ.
-        std::thread::sleep(std::time::Duration::from_millis(50));
-
-        let s2 = make_session("Second", vec![], vec![], vec![], vec![]);
-        let h2 = MeetingHandoff::from_session(&s2);
-        write_meeting_handoff(dir.path(), &h2).unwrap();
-
-        let loaded = load_meeting_handoff(dir.path()).unwrap().unwrap();
-        assert_eq!(loaded.topic, "Second");
-    }
-
-    #[test]
-    fn default_handoff_dir_respects_env() {
-        // This test is inherently environment-dependent — just verify it
-        // returns a non-empty path.
-        let dir = default_handoff_dir();
-        assert!(!dir.as_os_str().is_empty());
-    }
-
-    #[test]
-    fn from_session_explicit_questions_tagged() {
-        let mut session = make_session(
-            "Tagged",
-            vec!["What is the deadline for the release?"],
-            vec![],
-            vec![],
-            vec![],
-        );
-        session
-            .explicit_questions
-            .push("Who owns the rollback plan?".to_string());
-
-        let handoff = MeetingHandoff::from_session(&session);
-        assert_eq!(handoff.open_questions.len(), 2);
-
-        // Explicit question comes first.
-        assert_eq!(
-            handoff.open_questions[0].text,
-            "Who owns the rollback plan?"
-        );
-        assert!(handoff.open_questions[0].explicit);
-
-        // Inferred question from notes comes second.
-        assert_eq!(
-            handoff.open_questions[1].text,
-            "What is the deadline for the release?"
-        );
-        assert!(!handoff.open_questions[1].explicit);
-    }
-
-    #[test]
-    fn from_session_only_explicit_questions() {
-        let mut session = make_session("Explicit only", vec!["plain note"], vec![], vec![], vec![]);
-        session
-            .explicit_questions
-            .push("When do we ship?".to_string());
-
-        let handoff = MeetingHandoff::from_session(&session);
-        assert_eq!(handoff.open_questions.len(), 1);
-        assert_eq!(handoff.open_questions[0].text, "When do we ship?");
-        assert!(handoff.open_questions[0].explicit);
-    }
-
-    // -----------------------------------------------------------------------
-    // PR tests: find_newest_handoff direct tests + nonexistent dir
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn load_from_nonexistent_dir_returns_none() {
-        let result = load_meeting_handoff(Path::new("/nonexistent/path/for/testing")).unwrap();
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn find_newest_picks_latest_timestamped_file() {
-        let dir = tempfile::tempdir().unwrap();
-        // Write two files with different timestamps.
-        fs::write(
-            dir.path().join("handoff-2024-01-01T00-00-00_00-00.json"),
-            "{}",
-        )
-        .unwrap();
-        fs::write(
-            dir.path().join("handoff-2024-06-15T12-00-00_00-00.json"),
-            "{}",
-        )
-        .unwrap();
-        let newest = find_newest_handoff(dir.path()).unwrap();
-        assert!(
-            newest
-                .file_name()
-                .unwrap()
-                .to_string_lossy()
-                .contains("2024-06-15")
-        );
-    }
-
-    #[test]
-    fn find_newest_prefers_timestamped_over_legacy() {
-        let dir = tempfile::tempdir().unwrap();
-        fs::write(dir.path().join(MEETING_HANDOFF_FILENAME), "{}").unwrap();
-        fs::write(
-            dir.path().join("handoff-2099-01-01T00-00-00_00-00.json"),
-            "{}",
-        )
-        .unwrap();
-        let newest = find_newest_handoff(dir.path()).unwrap();
-        // The timestamped file sorts after "handoff-" prefix and legacy sorts
-        // after all timestamped files ("meeting_handoff.json" > "handoff-*"),
-        // so the legacy file is actually picked as "newest" by filename sort.
-        // Verify the function returns a valid path (either is acceptable as
-        // the function picks the last by lexicographic sort).
-        let name = newest.file_name().unwrap().to_string_lossy().to_string();
-        assert!(
-            name == MEETING_HANDOFF_FILENAME || name.contains("2099"),
-            "expected either legacy or timestamped file, got {name}"
-        );
-    }
-
-    // -----------------------------------------------------------------------
-    // WIP session persistence tests
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn save_and_load_wip_round_trip() {
-        let dir = tempfile::tempdir().unwrap();
-        let session = make_session(
-            "WIP test",
-            vec!["note one"],
-            vec![sample_decision()],
-            vec![sample_action()],
-            vec!["alice"],
-        );
-        save_session_wip(dir.path(), &session).unwrap();
-
-        let loaded = load_session_wip(dir.path()).unwrap().unwrap();
-        assert_eq!(loaded.topic, "WIP test");
-        assert_eq!(loaded.decisions.len(), 1);
-        assert_eq!(loaded.action_items.len(), 1);
-        assert_eq!(loaded.notes, vec!["note one"]);
-        assert_eq!(loaded.participants, vec!["alice"]);
-    }
-
-    #[test]
-    fn load_wip_returns_none_when_absent() {
-        let dir = tempfile::tempdir().unwrap();
-        let result = load_session_wip(dir.path()).unwrap();
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn remove_wip_deletes_file() {
-        let dir = tempfile::tempdir().unwrap();
-        let session = make_session("Remove me", vec![], vec![], vec![], vec![]);
-        save_session_wip(dir.path(), &session).unwrap();
-
-        // File should exist.
-        assert!(dir.path().join(MEETING_SESSION_WIP_FILENAME).is_file());
-
-        remove_session_wip(dir.path()).unwrap();
-
-        // File should be gone.
-        assert!(!dir.path().join(MEETING_SESSION_WIP_FILENAME).is_file());
-        // Loading should return None.
-        assert!(load_session_wip(dir.path()).unwrap().is_none());
-    }
-
-    #[test]
-    fn remove_wip_noop_when_absent() {
-        let dir = tempfile::tempdir().unwrap();
-        // Should not error when there's nothing to remove.
-        remove_session_wip(dir.path()).unwrap();
-    }
-
-    #[test]
-    fn save_wip_overwrites_previous() {
-        let dir = tempfile::tempdir().unwrap();
-        let s1 = make_session("First", vec![], vec![], vec![], vec![]);
-        save_session_wip(dir.path(), &s1).unwrap();
-
-        let s2 = make_session("Second", vec!["updated"], vec![], vec![], vec![]);
-        save_session_wip(dir.path(), &s2).unwrap();
-
-        let loaded = load_session_wip(dir.path()).unwrap().unwrap();
-        assert_eq!(loaded.topic, "Second");
-        assert_eq!(loaded.notes, vec!["updated"]);
-    }
 }
